+++ src/mykrobe/_vcf/__init__.py
@@ -1,1 +1,1 @@
-from mykrobe._vcf.models import VCF
+from mykrobe._vcf.models import VCF, Genotype